+++ flaml/automl/__init__.py
@@ -1,3 +1,5 @@
-from flaml.automl.automl import AutoML, AutoMLState, SearchState, logger_formatter, size
+from flaml.automl.automl import AutoML, size
+from flaml.automl.logger import logger_formatter
+from flaml.automl.state import SearchState, AutoMLState
 
 __all__ = ["AutoML", "AutoMLState", "SearchState", "logger_formatter", "size"]
