+++ flaml/automl/automl.py
@@ -6,30 +6,20 @@ from __future__ import annotations
 import time
 import os
 import sys
-from typing import Callable, Optional, List, Union, Any
+from typing import Callable, List, Union, Optional
 import inspect
 from functools import partial
 import numpy as np
-from scipy.sparse import issparse
-from sklearn.model_selection import (
-    train_test_split,
-    RepeatedStratifiedKFold,
-    RepeatedKFold,
-    GroupKFold,
-    TimeSeriesSplit,
-    GroupShuffleSplit,
-    StratifiedGroupKFold,
-)
-from sklearn.utils import shuffle
 from sklearn.base import BaseEstimator
 import pandas as pd
 import logging
 import json
+
+from flaml.automl.state import SearchState, AutoMLState
 from flaml.automl.ml import (
     compute_estimator,
     train_estimator,
     get_estimator_class,
-    get_classification_objective,
 )
 from flaml.config import (
     MIN_SAMPLE_TRAIN,
@@ -41,19 +31,13 @@ from flaml.config import (
     N_SPLITS,
     SAMPLE_MULTIPLY_FACTOR,
 )
-from flaml.automl.data import (
-    concat,
-    CLASSIFICATION,
-    TOKENCLASSIFICATION,
-    TS_FORECAST,
-    TS_FORECASTREGRESSION,
-    TS_FORECASTPANEL,
-    TS_TIMESTAMP_COL,
-    REGRESSION,
-    _is_nlp_task,
-    NLG_TASKS,
-)
+from flaml.automl.data import concat
+
+# TODO check to see when we can remove these
+from flaml.automl.task.task import CLASSIFICATION, TS_FORECAST, Task
+from flaml.automl.task.factory import task_factory
 from flaml import tune
+from flaml.automl.logger import logger, logger_formatter
 from flaml.automl.training_log import training_log_reader, training_log_writer
 from flaml.default import suggest_learner
 from flaml.version import __version__ as flaml_version
@@ -73,27 +57,6 @@ except ImportError:
         len_labels,
         unique_value_first_index,
     )
-
-logger = logging.getLogger(__name__)
-logger_formatter = logging.Formatter(
-    "[%(name)s: %(asctime)s] {%(lineno)d} %(levelname)s - %(message)s", "%m-%d %H:%M:%S"
-)
-logger.propagate = False
-
-try:
-    import mlflow
-except ImportError:
-    mlflow = None
-
-try:
-    from ray import __version__ as ray_version
-
-    assert ray_version >= "1.10.0"
-
-    ray_available = True
-except (ImportError, AssertionError):
-    ray_available = False
-
 try:
     os.environ["PYARROW_IGNORE_TIMEZONE"] = "1"
     import pyspark.pandas as ps
@@ -109,423 +72,19 @@ except ImportError:
         pass
 
 
-class SearchState:
-    @property
-    def search_space(self):
-        return self._search_space_domain
+try:
+    import mlflow
+except ImportError:
+    mlflow = None
 
-    @property
-    def estimated_cost4improvement(self):
-        return max(
-            self.time_best_found - self.time_best_found_old,
-            self.total_time_used - self.time_best_found,
-        )
+try:
+    from ray import __version__ as ray_version
 
-    def valid_starting_point_one_dim(self, value_one_dim, domain_one_dim):
-        from flaml.tune.space import sample
+    assert ray_version >= "1.10.0"
 
-        """
-            For each hp in the starting point, check the following 3 conditions:
-            (1) If the type of the starting point does not match the required type in search space, return false
-            (2) If the starting point is not in the required search space, return false
-            (3) If the search space is a value instead of domain, and the value is not equal to the starting point
-            Notice (2) include the case starting point not in user specified search space custom_hp
-        """
-        if isinstance(domain_one_dim, sample.Domain):
-            renamed_type = list(
-                inspect.signature(domain_one_dim.is_valid).parameters.values()
-            )[0].annotation
-            type_match = (
-                renamed_type == Any
-                or isinstance(value_one_dim, renamed_type)
-                or isinstance(value_one_dim, int)
-                and renamed_type is float
-            )
-            if not (type_match and domain_one_dim.is_valid(value_one_dim)):
-                return False
-        elif value_one_dim != domain_one_dim:
-            return False
-        return True
-
-    def valid_starting_point(self, starting_point, search_space):
-        return all(
-            self.valid_starting_point_one_dim(value, search_space[name].get("domain"))
-            for name, value in starting_point.items()
-            if name != "FLAML_sample_size"
-        )
-
-    def __init__(
-        self,
-        learner_class,
-        data_size,
-        task,
-        starting_point=None,
-        period=None,
-        custom_hp=None,
-        max_iter=None,
-        budget=None,
-    ):
-        self.init_eci = learner_class.cost_relative2lgbm() if budget >= 0 else 1
-        self._search_space_domain = {}
-        self.init_config = None
-        self.low_cost_partial_config = {}
-        self.cat_hp_cost = {}
-        self.data_size = data_size
-        self.ls_ever_converged = False
-        self.learner_class = learner_class
-        self._budget = budget
-        if task in TS_FORECAST:
-            search_space = learner_class.search_space(
-                data_size=data_size, task=task, pred_horizon=period
-            )
-        else:
-            search_space = learner_class.search_space(data_size=data_size, task=task)
-
-        if custom_hp is not None:
-            search_space.update(custom_hp)
-
-        if isinstance(starting_point, dict):
-            starting_point = AutoMLState.sanitize(starting_point)
-            if max_iter > 1 and not self.valid_starting_point(
-                starting_point, search_space
-            ):
-                # If the number of iterations is larger than 1, remove invalid point
-                logger.warning(
-                    "Starting point {} removed because it is outside of the search space".format(
-                        starting_point
-                    )
-                )
-                starting_point = None
-        elif isinstance(starting_point, list):
-            starting_point = [AutoMLState.sanitize(x) for x in starting_point]
-            if max_iter > len(starting_point):
-                # If the number of starting points is no smaller than max iter, avoid the checking
-                starting_point_len = len(starting_point)
-                starting_point = [
-                    x
-                    for x in starting_point
-                    if self.valid_starting_point(x, search_space)
-                ]
-                if starting_point_len > len(starting_point):
-                    logger.warning(
-                        "Starting points outside of the search space are removed. "
-                        f"Remaining starting points for {learner_class}: {starting_point}"
-                    )
-                starting_point = starting_point or None
-
-        for name, space in search_space.items():
-            assert (
-                "domain" in space
-            ), f"{name}'s domain is missing in the search space spec {space}"
-            if space["domain"] is None:
-                # don't search this hp
-                continue
-            self._search_space_domain[name] = space["domain"]
-
-            if "low_cost_init_value" in space:
-                self.low_cost_partial_config[name] = space["low_cost_init_value"]
-            if "cat_hp_cost" in space:
-                self.cat_hp_cost[name] = space["cat_hp_cost"]
-            # if a starting point is provided, set the init config to be
-            # the starting point provided
-            if (
-                isinstance(starting_point, dict)
-                and starting_point.get(name) is not None
-            ):
-                if self.init_config is None:
-                    self.init_config = {}
-                self.init_config[name] = starting_point[name]
-            elif (
-                not isinstance(starting_point, list)
-                and "init_value" in space
-                and self.valid_starting_point_one_dim(
-                    space["init_value"], space["domain"]
-                )
-            ):
-                if self.init_config is None:
-                    self.init_config = {}
-                self.init_config[name] = space["init_value"]
-
-        if isinstance(starting_point, list):
-            self.init_config = starting_point
-        else:
-            self.init_config = [] if self.init_config is None else [self.init_config]
-
-        self._hp_names = list(self._search_space_domain.keys())
-        self.search_alg = None
-        self.best_config = None
-        self.best_result = None
-        self.best_loss = self.best_loss_old = np.inf
-        self.total_time_used = 0
-        self.total_iter = 0
-        self.base_eci = None
-        self.time_best_found = self.time_best_found_old = 0
-        self.time2eval_best = 0
-        self.time2eval_best_old = 0
-        self.trained_estimator = None
-        self.sample_size = None
-        self.trial_time = 0
-
-    def update(self, result, time_used):
-        if result:
-            config = result["config"]
-            if config and "FLAML_sample_size" in config:
-                self.sample_size = config["FLAML_sample_size"]
-            else:
-                self.sample_size = self.data_size[0]
-            obj = result["val_loss"]
-            metric_for_logging = result["metric_for_logging"]
-            time2eval = result["time_total_s"]
-            trained_estimator = result["trained_estimator"]
-            del result["trained_estimator"]  # free up RAM
-            n_iter = (
-                trained_estimator
-                and hasattr(trained_estimator, "ITER_HP")
-                and trained_estimator.params.get(trained_estimator.ITER_HP)
-            )
-            if n_iter:
-                if "ml" in config:
-                    config["ml"][trained_estimator.ITER_HP] = n_iter
-                else:
-                    config[trained_estimator.ITER_HP] = n_iter
-        else:
-            obj, time2eval, trained_estimator = np.inf, 0.0, None
-            metric_for_logging = config = None
-        self.trial_time = time2eval
-        self.total_time_used += time_used if self._budget >= 0 else 1
-        self.total_iter += 1
-
-        if self.base_eci is None:
-            self.base_eci = time_used
-        if (obj is not None) and (obj < self.best_loss):
-            self.best_loss_old = self.best_loss if self.best_loss < np.inf else 2 * obj
-            self.best_loss = obj
-            self.best_result = result
-            self.time_best_found_old = self.time_best_found
-            self.time_best_found = self.total_time_used
-            self.iter_best_found = self.total_iter
-            self.best_config = config
-            self.best_config_sample_size = self.sample_size
-            self.best_config_train_time = time_used
-            if time2eval:
-                self.time2eval_best_old = self.time2eval_best
-                self.time2eval_best = time2eval
-            if (
-                self.trained_estimator
-                and trained_estimator
-                and self.trained_estimator != trained_estimator
-            ):
-                self.trained_estimator.cleanup()
-            if trained_estimator:
-                self.trained_estimator = trained_estimator
-        elif trained_estimator:
-            trained_estimator.cleanup()
-        self.metric_for_logging = metric_for_logging
-        self.val_loss, self.config = obj, config
-
-    def get_hist_config_sig(self, sample_size, config):
-        config_values = tuple([config[k] for k in self._hp_names if k in config])
-        config_sig = str(sample_size) + "_" + str(config_values)
-        return config_sig
-
-    def est_retrain_time(self, retrain_sample_size):
-        assert (
-            self.best_config_sample_size is not None
-        ), "need to first get best_config_sample_size"
-        return self.time2eval_best * retrain_sample_size / self.best_config_sample_size
-
-
-class AutoMLState:
-    def _prepare_sample_train_data(self, sample_size: int):
-        logger.debug("prepare_sample_train_data sample_size = {}".format(sample_size))
-        sampled_weight = groups = None
-        if sample_size <= self.data_size[0]:
-            if isinstance(self.X_train, (pd.DataFrame, psDataFrame)):
-                sampled_X_train = self.X_train.iloc[:sample_size]
-            else:
-                sampled_X_train = self.X_train[:sample_size]
-            if isinstance(self.y_train, (pd.Series, psSeries)):
-                sampled_y_train = self.y_train.iloc[:sample_size]
-            else:
-                sampled_y_train = self.y_train[:sample_size]
-            weight = self.fit_kwargs.get(
-                "sample_weight"
-            )  # NOTE: _prepare_sample_train_data is before kwargs is updated to fit_kwargs_by_estimator
-            if weight is not None:
-                sampled_weight = (
-                    weight.iloc[:sample_size]
-                    if isinstance(weight, (pd.Series, psSeries))
-                    else weight[:sample_size]
-                )
-            if self.groups is not None:
-                groups = (
-                    self.groups.iloc[:sample_size]
-                    if isinstance(self.groups, (pd.Series, psSeries))
-                    else self.groups[:sample_size]
-                )
-        else:
-            sampled_X_train = self.X_train_all
-            sampled_y_train = self.y_train_all
-            if (
-                "sample_weight" in self.fit_kwargs
-            ):  # NOTE: _prepare_sample_train_data is before kwargs is updated to fit_kwargs_by_estimator
-                sampled_weight = self.sample_weight_all
-            if self.groups is not None:
-                groups = self.groups_all
-        return sampled_X_train, sampled_y_train, sampled_weight, groups
-
-    @staticmethod
-    def _compute_with_config_base(
-        config_w_resource: dict,
-        state: AutoMLState,
-        estimator: str,
-        is_report: bool = True,
-    ) -> dict:
-        if "FLAML_sample_size" in config_w_resource:
-            sample_size = int(config_w_resource["FLAML_sample_size"])
-        else:
-            sample_size = state.data_size[0]
-
-        this_estimator_kwargs = state.fit_kwargs_by_estimator.get(
-            estimator
-        ).copy()  # NOTE: _compute_with_config_base is after kwargs is updated to fit_kwargs_by_estimator
-        (
-            sampled_X_train,
-            sampled_y_train,
-            sampled_weight,
-            groups,
-        ) = state._prepare_sample_train_data(sample_size)
-        if sampled_weight is not None:
-            weight = this_estimator_kwargs["sample_weight"]
-            this_estimator_kwargs["sample_weight"] = sampled_weight
-        if groups is not None:
-            this_estimator_kwargs["groups"] = groups
-        config = config_w_resource.copy()
-        if "FLAML_sample_size" in config:
-            del config["FLAML_sample_size"]
-        budget = (
-            None
-            if state.time_budget < 0
-            else state.time_budget - state.time_from_start
-            if sample_size == state.data_size[0]
-            else (state.time_budget - state.time_from_start)
-            / 2
-            * sample_size
-            / state.data_size[0]
-        )
-
-        (
-            trained_estimator,
-            val_loss,
-            metric_for_logging,
-            _,
-            pred_time,
-        ) = compute_estimator(
-            sampled_X_train,
-            sampled_y_train,
-            state.X_val,
-            state.y_val,
-            state.weight_val,
-            state.groups_val,
-            state.train_time_limit
-            if budget is None
-            else min(budget, state.train_time_limit or np.inf),
-            state.kf,
-            config,
-            state.task,
-            estimator,
-            state.eval_method,
-            state.metric,
-            state.best_loss,
-            state.n_jobs,
-            state.learner_classes.get(estimator),
-            state.cv_score_agg_func,
-            state.log_training_metric,
-            this_estimator_kwargs,
-            state.free_mem_ratio,
-        )
-        if state.retrain_final and not state.model_history:
-            trained_estimator.cleanup()
-
-        result = {
-            "pred_time": pred_time,
-            "wall_clock_time": time.time() - state._start_time_flag,
-            "metric_for_logging": metric_for_logging,
-            "val_loss": val_loss,
-            "trained_estimator": trained_estimator,
-        }
-        if sampled_weight is not None:
-            this_estimator_kwargs["sample_weight"] = weight
-        if is_report is True:
-            tune.report(**result)
-        return result
-
-    @classmethod
-    def sanitize(cls, config: dict) -> dict:
-        """Make a config ready for passing to estimator."""
-        config = config.get("ml", config).copy()
-        config.pop("FLAML_sample_size", None)
-        config.pop("learner", None)
-        config.pop("_choice_", None)
-        return config
-
-    def _train_with_config(
-        self,
-        estimator: str,
-        config_w_resource: dict,
-        sample_size: Optional[int] = None,
-    ):
-        if not sample_size:
-            sample_size = config_w_resource.get(
-                "FLAML_sample_size", len(self.y_train_all)
-            )
-        config = AutoMLState.sanitize(config_w_resource)
-
-        this_estimator_kwargs = self.fit_kwargs_by_estimator.get(
-            estimator
-        ).copy()  # NOTE: _train_with_config is after kwargs is updated to fit_kwargs_by_estimator
-        (
-            sampled_X_train,
-            sampled_y_train,
-            sampled_weight,
-            groups,
-        ) = self._prepare_sample_train_data(sample_size)
-        if sampled_weight is not None:
-            weight = this_estimator_kwargs[
-                "sample_weight"
-            ]  # NOTE: _train_with_config is after kwargs is updated to fit_kwargs_by_estimator
-            this_estimator_kwargs[
-                "sample_weight"
-            ] = sampled_weight  # NOTE: _train_with_config is after kwargs is updated to fit_kwargs_by_estimator
-        if groups is not None:
-            this_estimator_kwargs[
-                "groups"
-            ] = groups  # NOTE: _train_with_config is after kwargs is updated to fit_kwargs_by_estimator
-
-        budget = (
-            None if self.time_budget < 0 else self.time_budget - self.time_from_start
-        )
-
-        estimator, train_time = train_estimator(
-            X_train=sampled_X_train,
-            y_train=sampled_y_train,
-            config_dic=config,
-            task=self.task,
-            estimator_name=estimator,
-            n_jobs=self.n_jobs,
-            estimator_class=self.learner_classes.get(estimator),
-            budget=budget,
-            fit_kwargs=this_estimator_kwargs,  # NOTE: _train_with_config is after kwargs is updated to fit_kwargs_by_estimator
-            eval_metric=self.metric if hasattr(self, "metric") else "train_time",
-            free_mem_ratio=self.free_mem_ratio,
-        )
-
-        if sampled_weight is not None:
-            this_estimator_kwargs[
-                "sample_weight"
-            ] = weight  # NOTE: _train_with_config is after kwargs is updated to fit_kwargs_by_estimator
-
-        return estimator, train_time
+    ray_available = True
+except (ImportError, AssertionError):
+    ray_available = False
 
 
 def size(learner_classes: dict, config: dict) -> float:
@@ -609,7 +168,8 @@ class AutoML(BaseEstimator):
         ```
             task: A string of the task type, e.g.,
                 'classification', 'regression', 'ts_forecast', 'rank',
-                'seq-classification', 'seq-regression', 'summarization'.
+                'seq-classification', 'seq-regression', 'summarization',
+                or an instance of the Task class.
             n_jobs: An integer of the number of threads for training | default=-1.
                 Use all available resources when n_jobs == -1.
             log_file_name: A string of the log file name | default="". To disable logging,
@@ -991,7 +551,7 @@ class AutoML(BaseEstimator):
                 "No estimator is trained. Please run fit with enough budget."
             )
             return None
-        X = self._preprocess(X)
+        X = self._state.task.preprocess(X, self._transformer)
         if self._label_transformer:
             y = self._label_transformer.transform(y)
         return estimator.score(X, y, **kwargs)
@@ -1035,7 +595,7 @@ class AutoML(BaseEstimator):
                 "No estimator is trained. Please run fit with enough budget."
             )
             return None
-        X = self._preprocess(X)
+        X = self._state.task.preprocess(X, self._transformer)
         y_pred = estimator.predict(X, **pred_kwargs)
         if (
             isinstance(y_pred, np.ndarray)
@@ -1069,686 +629,9 @@ class AutoML(BaseEstimator):
                 "No estimator is trained. Please run fit with enough budget."
             )
             return None
-        X = self._preprocess(X)
+        X = self._state.task.preprocess(X, self._transformer)
         proba = self._trained_estimator.predict_proba(X, **pred_kwargs)
         return proba
-
-    def _preprocess(self, X):
-        if isinstance(X, List):
-            try:
-                if isinstance(X[0], List):
-                    X = [x for x in zip(*X)]
-                X = pd.DataFrame(
-                    dict(
-                        [
-                            (self._transformer._str_columns[idx], X[idx])
-                            if isinstance(X[0], List)
-                            else (self._transformer._str_columns[idx], [X[idx]])
-                            for idx in range(len(X))
-                        ]
-                    )
-                )
-            except IndexError:
-                raise IndexError(
-                    "Test data contains more columns than training data, exiting"
-                )
-        elif isinstance(X, int):
-            return X
-        elif isinstance(X, psDataFrame):
-            return X
-        elif issparse(X):
-            X = X.tocsr()
-        if self._state.task in TS_FORECAST:
-            X = pd.DataFrame(X)
-        if self._transformer:
-            X = self._transformer.transform(X)
-        return X
-
-    def _validate_ts_data(
-        self,
-        dataframe,
-        y_train_all=None,
-    ):
-        assert (
-            dataframe[dataframe.columns[0]].dtype.name == "datetime64[ns]"
-        ), f"For '{TS_FORECAST}' task, the first column must contain timestamp values."
-        if y_train_all is not None:
-            if isinstance(y_train_all, pd.Series):
-                y_df = pd.DataFrame(y_train_all)
-            elif isinstance(y_train_all, np.ndarray):
-                y_df = pd.DataFrame(y_train_all, columns=["labels"])
-            elif isinstance(y_train_all, (psDataFrame, psSeries)):
-                # TODO: optimize this
-                set_option("compute.ops_on_diff_frames", True)
-                y_df = y_train_all.to_frame()
-            dataframe = dataframe.join(y_df)
-        duplicates = dataframe.duplicated()
-        if isinstance(dataframe, psDataFrame):
-            if duplicates.any():
-                logger.warning("Duplicate timestamp values found in timestamp column.")
-                dataframe = dataframe.drop_duplicates()
-                logger.warning("Removed duplicate rows based on all columns")
-                assert (
-                    dataframe[[dataframe.columns[0]]].duplicated().any() is False
-                ), "Duplicate timestamp values with different values for other columns."
-                ts_series = ps.to_datetime(dataframe[dataframe.columns[0]])
-                inferred_freq = None  # `pd.infer_freq()` is not implemented yet.
-        else:
-            if any(duplicates):
-                logger.warning(
-                    "Duplicate timestamp values found in timestamp column. "
-                    f"\n{dataframe.loc[duplicates, dataframe][dataframe.columns[0]]}"
-                )
-                dataframe = dataframe.drop_duplicates()
-                logger.warning("Removed duplicate rows based on all columns")
-                assert (
-                    dataframe[[dataframe.columns[0]]].duplicated() is None
-                ), "Duplicate timestamp values with different values for other columns."
-            ts_series = pd.to_datetime(dataframe[dataframe.columns[0]])
-            inferred_freq = pd.infer_freq(ts_series)
-        if inferred_freq is None:
-            logger.warning(
-                "Missing timestamps detected. To avoid error with estimators, set estimator list to ['prophet']. "
-            )
-        if y_train_all is not None:
-            return dataframe.iloc[:, :-1], dataframe.iloc[:, -1]
-        return dataframe
-
-    def _validate_data(
-        self,
-        X_train_all,
-        y_train_all,
-        dataframe,
-        label,
-        X_val=None,
-        y_val=None,
-        groups_val=None,
-        groups=None,
-    ):
-        if X_train_all is not None and y_train_all is not None:
-            assert isinstance(
-                X_train_all, (np.ndarray, pd.DataFrame, psDataFrame)
-            ) or issparse(X_train_all), (
-                "X_train_all must be a numpy array, a pandas dataframe, "
-                "a Scipy sparse matrix or a pandas_on_spark dataframe."
-            )
-            assert isinstance(
-                y_train_all, (np.ndarray, pd.Series, psSeries)
-            ), "y_train_all must be a numpy array, a pandas series or a pandas_on_spark series."
-            assert (
-                X_train_all.size != 0 and y_train_all.size != 0
-            ), "Input data must not be empty."
-            if isinstance(X_train_all, np.ndarray) and len(X_train_all.shape) == 1:
-                X_train_all = np.reshape(X_train_all, (X_train_all.size, 1))
-            if isinstance(y_train_all, np.ndarray):
-                y_train_all = y_train_all.flatten()
-            assert (
-                X_train_all.shape[0] == y_train_all.shape[0]
-            ), "# rows in X_train must match length of y_train."
-            if isinstance(X_train_all, psDataFrame):
-                X_train_all = X_train_all.spark.cache()
-                shape = X_train_all.shape
-                logger.debug(f"X_train_all cached, shape of X_train_all: {shape}")
-            self._df = isinstance(X_train_all, (pd.DataFrame, psDataFrame))
-            self._nrow, self._ndim = X_train_all.shape
-            if self._state.task in TS_FORECAST:
-                X_train_all = (
-                    pd.DataFrame(X_train_all)
-                    if isinstance(X_train_all, np.ndarray)
-                    else X_train_all
-                )
-                X_train_all, y_train_all = self._validate_ts_data(
-                    X_train_all, y_train_all
-                )
-            X, y = X_train_all, y_train_all
-        elif dataframe is not None and label is not None:
-            assert isinstance(
-                dataframe, (pd.DataFrame, psDataFrame)
-            ), "dataframe must be a pandas DataFrame or a pandas_on_spark DataFrame."
-            assert (
-                label in dataframe.columns
-            ), f"The provided label column name `{label}` doesn't exist in the provided dataframe."
-            if isinstance(dataframe, psDataFrame):
-                dataframe = dataframe.spark.cache()
-                shape = dataframe.shape
-                logger.debug(f"dataframe cached, shape of dataframe: {shape}")
-            self._df = True
-            if self._state.task in TS_FORECAST:
-                dataframe = self._validate_ts_data(dataframe)
-            y = dataframe[label]
-            if isinstance(dataframe, pd.DataFrame):
-                X = dataframe.drop(columns=label)
-            else:
-                # TODO: to support pyspark.sql.DataFrame and pure dataframe mode in the following processes
-                X = dataframe
-                y.name = "y_" + y.name  # to avoid name conflict
-            self._nrow, self._ndim = X.shape
-        else:
-            raise ValueError("either X_train+y_train or dataframe+label are required")
-
-        # check the validity of input dimensions for NLP tasks, so need to check _is_nlp_task not estimator
-        if _is_nlp_task(self._state.task):
-            from .nlp.utils import is_a_list_of_str
-
-            is_all_str = True
-            is_all_list = True
-            for column in X.columns:
-                assert X[column].dtype.name in (
-                    "object",
-                    "string",
-                ), "If the task is an NLP task, X can only contain text columns"
-                for _, each_cell in X[column].items():
-                    if each_cell is not None:
-                        is_str = isinstance(each_cell, str)
-                        is_list_of_int = isinstance(each_cell, list) and all(
-                            isinstance(x, int) for x in each_cell
-                        )
-                        is_list_of_str = is_a_list_of_str(each_cell)
-                        if self._state.task == TOKENCLASSIFICATION:
-                            assert is_list_of_str, (
-                                "For the token-classification task, the input column needs to be a list of string,"
-                                "instead of string, e.g., ['EU', 'rejects','German', 'call','to','boycott','British','lamb','.',].",
-                                "For more examples, please refer to test/nlp/test_autohf_tokenclassification.py",
-                            )
-                        else:
-                            assert is_str or is_list_of_int, (
-                                "Each column of the input must either be str (untokenized) "
-                                "or a list of integers (tokenized)"
-                            )
-                        is_all_str &= is_str
-                        is_all_list &= is_list_of_int or is_list_of_str
-            assert is_all_str or is_all_list, (
-                "Currently FLAML only supports two modes for NLP: either all columns of X are string (non-tokenized), "
-                "or all columns of X are integer ids (tokenized)"
-            )
-
-        if isinstance(X, psDataFrame):
-            # TODO: support pyspark.pandas dataframe in DataTransformer
-            self._skip_transform = True
-
-        if self._skip_transform or issparse(X_train_all):
-            self._transformer = self._label_transformer = False
-            self._X_train_all, self._y_train_all = X, y
-        else:
-            from .data import DataTransformer
-
-            self._transformer = DataTransformer()
-
-            self._X_train_all, self._y_train_all = self._transformer.fit_transform(
-                X, y, self._state.task
-            )
-            self._label_transformer = self._transformer.label_transformer
-            if self._state.task == TOKENCLASSIFICATION:
-                if hasattr(self._label_transformer, "label_list"):
-                    self._state.fit_kwargs.update(
-                        {"label_list": self._label_transformer.label_list}
-                    )
-                elif "label_list" not in self._state.fit_kwargs:
-                    for each_fit_kwargs in self._state.fit_kwargs_by_estimator.values():
-                        assert (
-                            "label_list" in each_fit_kwargs
-                        ), "For the token-classification task, you must either (1) pass token labels; or (2) pass id labels and the label list. "
-                        "Please refer to the documentation for more details: https://microsoft.github.io/FLAML/docs/Examples/AutoML-NLP#a-simple-token-classification-example"
-            self._feature_names_in_ = (
-                self._X_train_all.columns.to_list()
-                if hasattr(self._X_train_all, "columns")
-                else None
-            )
-
-        self._sample_weight_full = self._state.fit_kwargs.get(
-            "sample_weight"
-        )  # NOTE: _validate_data is before kwargs is updated to fit_kwargs_by_estimator
-        if X_val is not None and y_val is not None:
-            assert isinstance(
-                X_val, (np.ndarray, pd.DataFrame, psDataFrame)
-            ) or issparse(X_train_all), (
-                "X_val must be None, a numpy array, a pandas dataframe, "
-                "a Scipy sparse matrix or a pandas_on_spark dataframe."
-            )
-            assert isinstance(y_val, (np.ndarray, pd.Series, psSeries)), (
-                "y_val must be None, a numpy array, a pandas series "
-                "or a pandas_on_spark series."
-            )
-            assert X_val.size != 0 and y_val.size != 0, (
-                "Validation data are expected to be nonempty. "
-                "Use None for X_val and y_val if no validation data."
-            )
-            if isinstance(y_val, np.ndarray):
-                y_val = y_val.flatten()
-            assert (
-                X_val.shape[0] == y_val.shape[0]
-            ), "# rows in X_val must match length of y_val."
-            if self._transformer:
-                self._state.X_val = self._transformer.transform(X_val)
-            else:
-                self._state.X_val = X_val
-            # If it's NLG_TASKS, y_val is a pandas series containing the output sequence tokens,
-            # so we cannot use label_transformer.transform to process it
-            if self._label_transformer:
-                self._state.y_val = self._label_transformer.transform(y_val)
-            else:
-                self._state.y_val = y_val
-        else:
-            self._state.X_val = self._state.y_val = None
-        if groups is not None and len(groups) != self._nrow:
-            # groups is given as group counts
-            self._state.groups = np.concatenate([[i] * c for i, c in enumerate(groups)])
-            assert (
-                len(self._state.groups) == self._nrow
-            ), "the sum of group counts must match the number of examples"
-            self._state.groups_val = (
-                np.concatenate([[i] * c for i, c in enumerate(groups_val)])
-                if groups_val is not None
-                else None
-            )
-        else:
-            self._state.groups_val = groups_val
-            self._state.groups = groups
-
-    def _split_pyspark(self, X_train_all, y_train_all, split_ratio, stratify=None):
-        # TODO: optimize this
-        set_option("compute.ops_on_diff_frames", True)
-        if not isinstance(y_train_all, (psDataFrame, psSeries)):
-            raise ValueError("y_train_all must be a pyspark.pandas dataframe or series")
-        df_all_in_one = X_train_all.join(y_train_all)
-        stratify_column = (
-            y_train_all.name
-            if isinstance(y_train_all, psSeries)
-            else y_train_all.columns[0]
-        )
-        ret_sample_weight = False
-        if (
-            "sample_weight" in self._state.fit_kwargs
-        ):  # NOTE: _prepare_data is before kwargs is updated to fit_kwargs_by_estimator
-            # fit_kwargs["sample_weight"] is an numpy array
-            ps_sample_weight = ps.DataFrame(
-                self._state.fit_kwargs["sample_weight"],
-                columns=["sample_weight"],
-            )
-            df_all_in_one = df_all_in_one.join(ps_sample_weight)
-            ret_sample_weight = True
-        df_all_train, df_all_val = train_test_split_pyspark(
-            df_all_in_one,
-            None if stratify is None else stratify_column,
-            test_fraction=split_ratio,
-            seed=RANDOM_SEED,
-        )
-        columns_to_drop = [
-            c for c in df_all_train.columns if c in [stratify_column, "sample_weight"]
-        ]
-        X_train = df_all_train.drop(columns_to_drop)
-        X_val = df_all_val.drop(columns_to_drop)
-        y_train = df_all_train[stratify_column]
-        y_val = df_all_val[stratify_column]
-
-        if ret_sample_weight:
-            return (
-                X_train,
-                X_val,
-                y_train,
-                y_val,
-                df_all_train["sample_weight"].to_numpy(),
-                df_all_val["sample_weight"].to_numpy(),
-            )
-        return X_train, X_val, y_train, y_val
-
-    def _train_test_split(
-        self, X, y, first=None, rest=None, split_ratio=0.2, stratify=None
-    ):
-        condition_type = isinstance(X, (psDataFrame, psSeries))
-        # NOTE: _prepare_data is before kwargs is updated to fit_kwargs_by_estimator
-        condition_param = "sample_weight" in self._state.fit_kwargs
-        if not condition_type and condition_param:
-            sample_weight = (
-                self._state.fit_kwargs["sample_weight"]
-                if rest is None
-                else self._state.fit_kwargs["sample_weight"][rest]
-            )
-            (
-                X_train,
-                X_val,
-                y_train,
-                y_val,
-                weight_train,
-                weight_val,
-            ) = train_test_split(
-                X,
-                y,
-                sample_weight,
-                test_size=split_ratio,
-                stratify=stratify,
-                random_state=RANDOM_SEED,
-            )
-
-            if first is not None:
-                weight1 = self._state.fit_kwargs["sample_weight"][first]
-                self._state.weight_val = concat(weight1, weight_val)
-                self._state.fit_kwargs["sample_weight"] = concat(weight1, weight_train)
-            else:
-                self._state.weight_val = weight_val
-                self._state.fit_kwargs["sample_weight"] = weight_train
-        elif not condition_type and not condition_param:
-            X_train, X_val, y_train, y_val = train_test_split(
-                X,
-                y,
-                test_size=split_ratio,
-                stratify=stratify,
-                random_state=RANDOM_SEED,
-            )
-        elif condition_type and condition_param:
-            (
-                X_train,
-                X_val,
-                y_train,
-                y_val,
-                weight_train,
-                weight_val,
-            ) = self._split_pyspark(X, y, split_ratio, stratify)
-
-            if first is not None:
-                weight1 = self._state.fit_kwargs["sample_weight"][first]
-                self._state.weight_val = concat(weight1, weight_val)
-                self._state.fit_kwargs["sample_weight"] = concat(weight1, weight_train)
-            else:
-                self._state.weight_val = weight_val
-                self._state.fit_kwargs["sample_weight"] = weight_train
-        else:
-            X_train, X_val, y_train, y_val = self._split_pyspark(
-                X, y, split_ratio, stratify
-            )
-        return X_train, X_val, y_train, y_val
-
-    def _prepare_data(self, eval_method, split_ratio, n_splits):
-        X_val, y_val = self._state.X_val, self._state.y_val
-        if issparse(X_val):
-            X_val = X_val.tocsr()
-        X_train_all, y_train_all = self._X_train_all, self._y_train_all
-        if issparse(X_train_all):
-            X_train_all = X_train_all.tocsr()
-        if (
-            self._state.task in CLASSIFICATION
-            and self._auto_augment
-            and self._state.fit_kwargs.get("sample_weight")
-            is None  # NOTE: _prepare_data is before kwargs is updated to fit_kwargs_by_estimator
-            and self._split_type in ["stratified", "uniform"]
-            and self._state.task != TOKENCLASSIFICATION
-        ):
-            # logger.info(f"label {pd.unique(y_train_all)}")
-            if not isinstance(y_train_all, (psDataFrame, psSeries)):
-                label_set, counts = np.unique(y_train_all, return_counts=True)
-            else:
-                label_set, counts = unique_pandas_on_spark(y_train_all)
-            # augment rare classes
-            rare_threshld = 20
-            rare = counts < rare_threshld
-            rare_label, rare_counts = label_set[rare], counts[rare]
-            for i, label in enumerate(rare_label.tolist()):
-                count = rare_count = rare_counts[i]
-                rare_index = y_train_all == label
-                n = len(y_train_all)
-                while count < rare_threshld:
-                    if isinstance(X_train_all, (psDataFrame, psSeries)):
-                        # TODO: optimize this
-                        set_option("compute.ops_on_diff_frames", True)
-                    if self._df:
-                        X_train_all = concat(
-                            X_train_all, X_train_all.iloc[:n].loc[rare_index]
-                        )
-                    else:
-                        X_train_all = concat(
-                            X_train_all, X_train_all[:n][rare_index, :]
-                        )
-                    if isinstance(y_train_all, (pd.Series, psSeries)):
-                        y_train_all = concat(
-                            y_train_all, y_train_all.iloc[:n].loc[rare_index]
-                        )
-                    else:
-                        y_train_all = np.concatenate(
-                            [y_train_all, y_train_all[:n][rare_index]]
-                        )
-                    count += rare_count
-                logger.info(f"class {label} augmented from {rare_count} to {count}")
-        SHUFFLE_SPLIT_TYPES = ["uniform", "stratified"]
-        if isinstance(y_train_all, (psDataFrame, psSeries)):
-            # no need to shuffle pyspark dataframe
-            pass
-        elif self._split_type in SHUFFLE_SPLIT_TYPES:
-            if self._sample_weight_full is not None:
-                X_train_all, y_train_all, self._state.sample_weight_all = shuffle(
-                    X_train_all,
-                    y_train_all,
-                    self._sample_weight_full,
-                    random_state=RANDOM_SEED,
-                )
-                self._state.fit_kwargs[
-                    "sample_weight"
-                ] = (
-                    self._state.sample_weight_all
-                )  # NOTE: _prepare_data is before kwargs is updated to fit_kwargs_by_estimator
-                if isinstance(self._state.sample_weight_all, pd.Series):
-                    self._state.sample_weight_all.reset_index(drop=True, inplace=True)
-            else:
-                X_train_all, y_train_all = shuffle(
-                    X_train_all, y_train_all, random_state=RANDOM_SEED
-                )
-            if self._df:
-                X_train_all.reset_index(drop=True, inplace=True)
-            if isinstance(y_train_all, pd.Series):
-                y_train_all.reset_index(drop=True, inplace=True)
-
-        X_train, y_train = X_train_all, y_train_all
-        self._state.groups_all = self._state.groups
-        if X_val is None and eval_method == "holdout":
-            # if eval_method = holdout, make holdout data
-            if self._split_type == "time":
-                if self._state.task in TS_FORECAST:
-                    period = self._state.fit_kwargs[
-                        "period"
-                    ]  # NOTE: _prepare_data is before kwargs is updated to fit_kwargs_by_estimator
-                    if self._state.task == TS_FORECASTPANEL:
-                        X_train_all["time_idx"] -= X_train_all["time_idx"].min()
-                        X_train_all["time_idx"] = X_train_all["time_idx"].astype("int")
-                        ids = self._state.fit_kwargs["group_ids"].copy()
-                        ids.append(TS_TIMESTAMP_COL)
-                        ids.append("time_idx")
-                        y_train_all = (
-                            pd.DataFrame(y_train_all)
-                            if not isinstance(y_train_all, (psDataFrame, psSeries))
-                            else ps.DataFrame(y_train_all)
-                            if isinstance(y_train_all, psSeries)
-                            else y_train_all
-                        )
-                        y_train_all[ids] = X_train_all[ids]
-                        X_train_all = X_train_all.sort_values(ids)
-                        y_train_all = y_train_all.sort_values(ids)
-                        training_cutoff = X_train_all["time_idx"].max() - period
-                        X_train = X_train_all[
-                            X_train_all["time_idx"] <= training_cutoff
-                        ]
-                        y_train = y_train_all[
-                            y_train_all["time_idx"] <= training_cutoff
-                        ].drop(columns=ids)
-                        X_val = X_train_all[X_train_all["time_idx"] > training_cutoff]
-                        y_val = y_train_all[
-                            y_train_all["time_idx"] > training_cutoff
-                        ].drop(columns=ids)
-                    else:
-                        num_samples = X_train_all.shape[0]
-                        assert (
-                            period < num_samples
-                        ), f"period={period}>#examples={num_samples}"
-                        split_idx = num_samples - period
-                        X_train = X_train_all[:split_idx]
-                        y_train = y_train_all[:split_idx]
-                        X_val = X_train_all[split_idx:]
-                        y_val = y_train_all[split_idx:]
-                else:
-                    condition_type = isinstance(X_train_all, (psDataFrame, psSeries))
-                    # NOTE: _prepare_data is before kwargs is updated to fit_kwargs_by_estimator
-                    condition_param = "sample_weight" in self._state.fit_kwargs
-                    if not condition_type and condition_param:
-                        (
-                            X_train,
-                            X_val,
-                            y_train,
-                            y_val,
-                            self._state.fit_kwargs[
-                                "sample_weight"
-                            ],  # NOTE: _prepare_data is before kwargs is updated to fit_kwargs_by_estimator
-                            self._state.weight_val,
-                        ) = train_test_split(
-                            X_train_all,
-                            y_train_all,
-                            self._state.fit_kwargs[
-                                "sample_weight"
-                            ],  # NOTE: _prepare_data is before kwargs is updated to fit_kwargs_by_estimator
-                            test_size=split_ratio,
-                            shuffle=False,
-                        )
-                    elif not condition_type and not condition_param:
-                        X_train, X_val, y_train, y_val = train_test_split(
-                            X_train_all,
-                            y_train_all,
-                            test_size=split_ratio,
-                            shuffle=False,
-                        )
-                    elif condition_type and condition_param:
-                        (
-                            X_train,
-                            X_val,
-                            y_train,
-                            y_val,
-                            self._state.fit_kwargs[
-                                "sample_weight"
-                            ],  # NOTE: _prepare_data is before kwargs is updated to fit_kwargs_by_estimator
-                            self._state.weight_val,
-                        ) = self._split_pyspark(X_train_all, y_train_all, split_ratio)
-                    else:
-                        X_train, X_val, y_train, y_val = self._split_pyspark(
-                            X_train_all, y_train_all, split_ratio
-                        )
-            elif self._split_type == "group":
-                gss = GroupShuffleSplit(
-                    n_splits=1, test_size=split_ratio, random_state=RANDOM_SEED
-                )
-                for train_idx, val_idx in gss.split(
-                    X_train_all, y_train_all, self._state.groups_all
-                ):
-                    if self._df:
-                        X_train = X_train_all.iloc[train_idx]
-                        X_val = X_train_all.iloc[val_idx]
-                    else:
-                        X_train, X_val = X_train_all[train_idx], X_train_all[val_idx]
-                    y_train, y_val = y_train_all[train_idx], y_train_all[val_idx]
-                    self._state.groups = self._state.groups_all[train_idx]
-                    self._state.groups_val = self._state.groups_all[val_idx]
-            elif self._state.task in CLASSIFICATION:
-                # for classification, make sure the labels are complete in both
-                # training and validation data
-                label_set, first = unique_value_first_index(y_train_all)
-                rest = []
-                last = 0
-                first.sort()
-                for i in range(len(first)):
-                    rest.extend(range(last, first[i]))
-                    last = first[i] + 1
-                rest.extend(range(last, len(y_train_all)))
-                X_first = X_train_all.iloc[first] if self._df else X_train_all[first]
-                X_rest = X_train_all.iloc[rest] if self._df else X_train_all[rest]
-                y_rest = (
-                    y_train_all[rest]
-                    if isinstance(y_train_all, np.ndarray)
-                    else y_train_all.iloc[rest]
-                )
-                stratify = y_rest if self._split_type == "stratified" else None
-                X_train, X_val, y_train, y_val = self._train_test_split(
-                    X_rest, y_rest, first, rest, split_ratio, stratify
-                )
-                X_train = concat(X_first, X_train)
-                y_train = (
-                    concat(label_set, y_train)
-                    if self._df
-                    else np.concatenate([label_set, y_train])
-                )
-                X_val = concat(X_first, X_val)
-                y_val = (
-                    concat(label_set, y_val)
-                    if self._df
-                    else np.concatenate([label_set, y_val])
-                )
-            elif self._state.task in REGRESSION:
-                X_train, X_val, y_train, y_val = self._train_test_split(
-                    X_train_all, y_train_all, split_ratio=split_ratio
-                )
-        self._state.data_size = X_train.shape
-        self.data_size_full = len(y_train_all)
-        self._state.X_train, self._state.y_train = X_train, y_train
-        self._state.X_val, self._state.y_val = X_val, y_val
-        self._state.X_train_all = X_train_all
-        self._state.y_train_all = y_train_all
-        if eval_method == "holdout":
-            self._state.kf = None
-            return
-        if self._split_type == "group":
-            # logger.info("Using GroupKFold")
-            assert (
-                len(self._state.groups_all) == y_train_all.size
-            ), "the length of groups must match the number of examples"
-            assert (
-                len(np.unique(self._state.groups_all)) >= n_splits
-            ), "the number of groups must be equal or larger than n_splits"
-            self._state.kf = GroupKFold(n_splits)
-        elif self._split_type == "stratified":
-            # logger.info("Using StratifiedKFold")
-            assert self.data_size_full >= n_splits, (
-                f"{n_splits}-fold cross validation"
-                f" requires input data with at least {n_splits} examples."
-            )
-            assert self.data_size_full >= 2 * n_splits, (
-                f"{n_splits}-fold cross validation with metric=r2 "
-                f"requires input data with at least {n_splits*2} examples."
-            )
-            self._state.kf = RepeatedStratifiedKFold(
-                n_splits=n_splits, n_repeats=1, random_state=RANDOM_SEED
-            )
-        elif self._split_type == "time":
-            # logger.info("Using TimeSeriesSplit")
-            if (
-                self._state.task in TS_FORECAST
-                and self._state.task is not TS_FORECASTPANEL
-            ):
-                period = self._state.fit_kwargs[
-                    "period"
-                ]  # NOTE: _prepare_data is before kwargs is updated to fit_kwargs_by_estimator
-                if period * (n_splits + 1) > y_train_all.size:
-                    n_splits = int(y_train_all.size / period - 1)
-                    assert n_splits >= 2, (
-                        f"cross validation for forecasting period={period}"
-                        f" requires input data with at least {3 * period} examples."
-                    )
-                    logger.info(f"Using nsplits={n_splits} due to data size limit.")
-                self._state.kf = TimeSeriesSplit(n_splits=n_splits, test_size=period)
-            elif self._state.task is TS_FORECASTPANEL:
-                n_groups = len(
-                    X_train.groupby(self._state.fit_kwargs.get("group_ids")).size()
-                )
-                period = self._state.fit_kwargs.get("period")
-                self._state.kf = TimeSeriesSplit(
-                    n_splits=n_splits, test_size=period * n_groups
-                )
-            else:
-                self._state.kf = TimeSeriesSplit(n_splits=n_splits)
-        elif isinstance(self._split_type, str):
-            # logger.info("Using RepeatedKFold")
-            self._state.kf = RepeatedKFold(
-                n_splits=n_splits, n_repeats=1, random_state=RANDOM_SEED
-            )
-        else:
-            # logger.info("Using splitter object")
-            self._state.kf = self._split_type
-        if isinstance(self._state.kf, (GroupKFold, StratifiedGroupKFold)):
-            # self._split_type is either "group", a GroupKFold object, or a StratifiedGroupKFold object
-            self._state.kf.groups = self._state.groups_all
 
     def add_learner(self, learner_name, learner_class):
         """Add a customized learner.
@@ -1759,7 +642,9 @@ class AutoML(BaseEstimator):
         """
         self._state.learner_classes[learner_name] = learner_class
 
-    def get_estimator_from_log(self, log_file_name: str, record_id: int, task: str):
+    def get_estimator_from_log(
+        self, log_file_name: str, record_id: int, task: Union[str, Task]
+    ):
         """Get the estimator from log file.
 
         Args:
@@ -1767,7 +652,8 @@ class AutoML(BaseEstimator):
             record_id: An integer of the record ID in the file,
                 0 corresponds to the first trial.
             task: A string of the task type,
-                'binary', 'multiclass', 'regression', 'ts_forecast', 'rank'.
+                'binary', 'multiclass', 'regression', 'ts_forecast', 'rank',
+                or an instance of the Task class.
 
         Returns:
             An estimator object for the given configuration.
@@ -1777,6 +663,9 @@ class AutoML(BaseEstimator):
             record = reader.get_record(record_id)
             estimator = record.learner
             config = AutoMLState.sanitize(record.config)
+
+        if isinstance(task, str):
+            task = task_factory(task)
 
         estimator, _ = train_estimator(
             X_train=None,
@@ -1797,7 +686,7 @@ class AutoML(BaseEstimator):
         dataframe=None,
         label=None,
         time_budget=np.inf,
-        task=None,
+        task: Optional[Union[str, Task]] = None,
         eval_method=None,
         split_ratio=None,
         n_splits=None,
@@ -1824,10 +713,7 @@ class AutoML(BaseEstimator):
         Args:
             log_file_name: A string of the log file name.
             X_train: A numpy array or dataframe of training data in shape n*m.
-                For time series forecast tasks, the first column of X_train
-                must be the timestamp column (datetime type). Other
-                columns in the dataframe are assumed to be exogenous
-                variables (categorical or numeric).
+                For time series forecast tasks, the first column of X_train must be the timestamp column (datetime type). Other columns in the dataframe are assumed to be exogenous variables (categorical or numeric).
             y_train: A numpy array or series of labels in shape n*1.
             dataframe: A dataframe of training data including label column.
                 For time series forecast tasks, dataframe must be specified and should
@@ -1842,7 +728,8 @@ class AutoML(BaseEstimator):
             time_budget: A float number of the time budget in seconds.
             task: A string of the task type, e.g.,
                 'classification', 'regression', 'ts_forecast', 'rank',
-                'seq-classification', 'seq-regression', 'summarization'.
+                'seq-classification', 'seq-regression', 'summarization',
+                or an instance of Task class.
             eval_method: A string of resampling strategy, one of
                 ['auto', 'cv', 'holdout'].
             split_ratio: A float of the validation data percentage for holdout.
@@ -1928,6 +815,9 @@ class AutoML(BaseEstimator):
                         used by TemporalFusionTransformerEstimator.
         """
         task = task or self._settings.get("task")
+        if isinstance(task, str):
+            task = task_factory(task)
+
         eval_method = eval_method or self._settings.get("eval_method")
         split_ratio = split_ratio or self._settings.get("split_ratio")
         n_splits = n_splits or self._settings.get("n_splits")
@@ -1936,7 +826,7 @@ class AutoML(BaseEstimator):
             self._settings.get("auto_augment") if auto_augment is None else auto_augment
         )
         self._state.task = task
-        self._estimator_type = "classifier" if task in CLASSIFICATION else "regressor"
+        self._estimator_type = "classifier" if task.is_classification() else "regressor"
 
         self._state.fit_kwargs = fit_kwargs
         self._state.custom_hp = custom_hp or self._settings.get("custom_hp")
@@ -1953,7 +843,9 @@ class AutoML(BaseEstimator):
             if preserve_checkpoint is None
             else preserve_checkpoint
         )
-        self._validate_data(X_train, y_train, dataframe, label, groups=groups)
+        task.validate_data(
+            self, self._state, X_train, y_train, dataframe, label, groups=groups
+        )
 
         logger.info("log file name {}".format(log_file_name))
 
@@ -2017,7 +909,12 @@ class AutoML(BaseEstimator):
         )
         # Partially copied from fit() function
         # Initilize some attributes required for retrain_from_log
-        self._decide_split_type(split_type)
+        self._split_type = task.decide_split_type(
+            split_type,
+            self._y_train_all,
+            self._state.fit_kwargs,
+            self._state.groups,
+        )
         eval_method = self._decide_eval_method(eval_method, time_budget)
         self.modelcount = 0
         self._auto_augment = auto_augment
@@ -2042,50 +939,6 @@ class AutoML(BaseEstimator):
         )[0]
         logger.info("retrain from log succeeded")
         return training_duration
-
-    def _decide_split_type(self, split_type):
-        if self._state.task == "classification":
-            num_classes = len_labels(self._y_train_all)
-            self._state.task = get_classification_objective(num_classes)
-        if not isinstance(split_type, str):
-            assert hasattr(split_type, "split") and hasattr(
-                split_type, "get_n_splits"
-            ), "split_type must be a string or a splitter object with split and get_n_splits methods."
-            assert (
-                not isinstance(split_type, GroupKFold) or self._state.groups is not None
-            ), "GroupKFold requires groups to be provided."
-            self._split_type = split_type
-        elif self._state.task in CLASSIFICATION:
-            assert split_type in ["auto", "stratified", "uniform", "time", "group"]
-            self._split_type = (
-                split_type
-                if split_type != "auto"
-                else self._state.groups is None and "stratified" or "group"
-            )
-        elif self._state.task in REGRESSION:
-            assert split_type in ["auto", "uniform", "time", "group"]
-            self._split_type = split_type if split_type != "auto" else "uniform"
-        elif self._state.task in TS_FORECAST:
-            assert split_type in ["auto", "time"]
-            self._split_type = "time"
-            assert isinstance(
-                self._state.fit_kwargs.get("period"),
-                int,  # NOTE: _decide_split_type is before kwargs is updated to fit_kwargs_by_estimator
-            ), f"missing a required integer 'period' for '{TS_FORECAST}' task."
-            if self._state.fit_kwargs.get("group_ids"):
-                self._state.task == TS_FORECASTPANEL
-                assert isinstance(
-                    self._state.fit_kwargs.get("group_ids"), list
-                ), f"missing a required List[str] 'group_ids' for '{TS_FORECASTPANEL}' task."
-        elif self._state.task == "rank":
-            assert (
-                self._state.groups is not None
-            ), "groups must be specified for ranking task."
-            assert split_type in ["auto", "group"]
-            self._split_type = "group"
-        elif self._state.task in NLG_TASKS:
-            assert split_type in ["auto", "uniform", "time", "group"]
-            self._split_type = split_type if split_type != "auto" else "uniform"
 
     def _decide_eval_method(self, eval_method, time_budget):
         if not isinstance(self._split_type, str):
@@ -2334,6 +1187,21 @@ class AutoML(BaseEstimator):
         """
         return self._metric_constraints
 
+    def _prepare_data(self, eval_method, split_ratio, n_splits):
+        self._state.task.prepare_data(
+            self._state,
+            self._X_train_all,
+            self._y_train_all,
+            self._auto_augment,
+            eval_method,
+            self._split_type,
+            split_ratio,
+            n_splits,
+            self._df,
+            self._sample_weight_full,
+        )
+        self.data_size_full = len(self._state.y_train_all)
+
     def fit(
         self,
         X_train=None,
@@ -2341,7 +1209,7 @@ class AutoML(BaseEstimator):
         dataframe=None,
         label=None,
         metric=None,
-        task=None,
+        task: Optional[Union[str, Task]] = None,
         n_jobs=None,
         # gpu_per_trial=0,
         log_file_name=None,
@@ -2451,7 +1319,7 @@ class AutoML(BaseEstimator):
             task: A string of the task type, e.g.,
                 'classification', 'regression', 'ts_forecast_regression',
                 'ts_forecast_classification', 'rank', 'seq-classification',
-                'seq-regression', 'summarization'.
+                'seq-regression', 'summarization', or an instance of Task class
             n_jobs: An integer of the number of threads for training | default=-1.
                 Use all available resources when n_jobs == -1.
             log_file_name: A string of the log file name | default="". To disable logging,
@@ -2701,7 +1569,10 @@ class AutoML(BaseEstimator):
 
         self._state._start_time_flag = self._start_time_flag = time.time()
         task = task or self._settings.get("task")
-        self._estimator_type = "classifier" if task in CLASSIFICATION else "regressor"
+        if isinstance(task, str):
+            task = task_factory(task, X_train, y_train)
+        self._state.task = task
+        self._estimator_type = "classifier" if task.is_classification() else "regressor"
         time_budget = time_budget or self._settings.get("time_budget")
         n_jobs = n_jobs or self._settings.get("n_jobs")
         gpu_per_trial = fit_kwargs.get("gpu_per_trial", 0)
@@ -2881,15 +1752,29 @@ class AutoML(BaseEstimator):
         )  # shallow copy of fit_kwargs_by_estimator
         self._state.weight_val = sample_weight_val
 
-        self._validate_data(
-            X_train, y_train, dataframe, label, X_val, y_val, groups_val, groups
+        task.validate_data(
+            self,
+            self._state,
+            X_train,
+            y_train,
+            dataframe,
+            label,
+            X_val,
+            y_val,
+            groups_val,
+            groups,
         )
         self._search_states = {}  # key: estimator name; value: SearchState
         self._random = np.random.RandomState(RANDOM_SEED)
         self._seed = seed if seed is not None else 20
         self._learner_selector = learner_selector
         logger.info(f"task = {task}")
-        self._decide_split_type(split_type)
+        self._split_type = self._state.task.decide_split_type(
+            split_type,
+            self._y_train_all,
+            self._state.fit_kwargs,
+            self._state.groups,
+        )
         logger.info(f"Data split method: {self._split_type}")
         eval_method = self._decide_eval_method(eval_method, time_budget)
         self._state.eval_method = eval_method
@@ -2940,12 +1825,13 @@ class AutoML(BaseEstimator):
         self._min_sample_size_input = min_sample_size
         self._prepare_data(eval_method, split_ratio, n_splits)
 
+        # TODO pull this to task as decide_sample_size
         if isinstance(self._min_sample_size, dict):
             self._sample = {
                 (
                     k,
                     sample
-                    and task != "rank"
+                    and not task.is_rank()
                     and eval_method != "cv"
                     and (
                         self._min_sample_size[k] * SAMPLE_MULTIPLY_FACTOR
@@ -2957,33 +1843,18 @@ class AutoML(BaseEstimator):
         else:
             self._sample = (
                 sample
-                and task != "rank"
+                and not task.is_rank()
                 and eval_method != "cv"
                 and (
                     self._min_sample_size * SAMPLE_MULTIPLY_FACTOR
                     < self._state.data_size[0]
                 )
             )
-        if "auto" == metric:
-            if _is_nlp_task(self._state.task):
-                from flaml.automl.nlp.utils import (
-                    load_default_huggingface_metric_for_task,
-                )
 
-                metric = load_default_huggingface_metric_for_task(self._state.task)
-            elif "binary" in self._state.task:
-                metric = "roc_auc"
-            elif "multiclass" in self._state.task:
-                metric = "log_loss"
-            elif self._state.task in TS_FORECAST:
-                metric = "mape"
-            elif self._state.task == "rank":
-                metric = "ndcg"
-            else:
-                metric = "r2"
-
+        metric = task.default_metric(metric)
         self._state.metric = metric
 
+        # TODO pull this to task
         def is_to_reverse_metric(metric, task):
             if metric.startswith("ndcg"):
                 return True, f"1-{metric}"
@@ -3002,7 +1873,7 @@ class AutoML(BaseEstimator):
                 "macro_f1",
             ]:
                 return True, f"1-{metric}"
-            if _is_nlp_task(task):
+            if task.is_nlp():
                 from flaml.automl.ml import huggingface_metric_to_mode
 
                 if (
@@ -3022,77 +1893,53 @@ class AutoML(BaseEstimator):
             error_metric = "customized metric"
         logger.info(f"Minimizing error metric: {error_metric}")
 
-        if "auto" == estimator_list:
-            # TODO: update estimator_list for spark dataframe
-            if isinstance(X_train, psDataFrame) or isinstance(dataframe, psDataFrame):
-                if self._state.task in ["rank", "classification", "regression"]:
-                    estimator_list = ["lgbm_spark"]
-                else:
-                    raise ValueError(
-                        "Spark dataframe is currently supported for classification, regression, and rank tasks."
-                        f"For {self._state.task} task, please convert the dataframe to pandas dataframe."
-                    )
-            elif self._state.task == "rank":
-                estimator_list = ["lgbm", "xgboost", "xgb_limitdepth"]
-            elif _is_nlp_task(self._state.task):
-                estimator_list = ["transformer"]
-            elif self._state.task == TS_FORECASTPANEL:
-                estimator_list = ["tft"]
-            else:
-                try:
-                    import catboost
-
-                    estimator_list = [
-                        "lgbm",
-                        "rf",
-                        "catboost",
-                        "xgboost",
-                        "extra_tree",
-                        "xgb_limitdepth",
-                    ]
-                except ImportError:
-                    estimator_list = [
-                        "lgbm",
-                        "rf",
-                        "xgboost",
-                        "extra_tree",
-                        "xgb_limitdepth",
-                    ]
-                if self._state.task in TS_FORECAST:
-                    # catboost is removed because it has a `name` parameter, making it incompatible with hcrystalball
-                    if "catboost" in estimator_list:
-                        estimator_list.remove("catboost")
-                    if self._state.task in TS_FORECASTREGRESSION:
-                        try:
-                            import prophet
-
-                            estimator_list += ["prophet", "arima", "sarimax"]
-                        except ImportError:
-                            estimator_list += ["arima", "sarimax"]
-                elif "regression" != self._state.task:
-                    estimator_list += ["lrl1"]
-        # For spark dataframe, only estimators ends with '_spark' are supported
-        # and use_spark must be False because spark models are trained in parallel themselves
-        if isinstance(X_train, psDataFrame) or isinstance(dataframe, psDataFrame):
-            if self._use_spark:
-                self._use_spark = False
-                logger.warning(
-                    "Spark dataframes support only spark.ml type models, which will be trained "
-                    "with spark themselves, no need to start spark trials in flaml. "
-                    "`use_spark` is set to False."
-                )
+        is_auto = estimator_list == "auto"
+        is_spark_dataframe = isinstance(X_train, psDataFrame) or isinstance(
+            dataframe, psDataFrame
+        )
+        estimator_list = task.default_estimator_list(estimator_list)
+        if not is_auto:
             n_estimators = len(estimator_list)
-            estimator_list = [est for est in estimator_list if est.endswith("_spark")]
-            if len(estimator_list) == 0:
-                raise ValueError(
-                    "Spark dataframes only support estimator names ending with `_spark`. Non-supported "
-                    "estimators are removed. No estimator is left."
-                )
-            elif n_estimators != len(estimator_list):
-                logger.warning(
-                    "Spark dataframes only support estimator names ending with `_spark`. Non-supported "
-                    "estimators are removed."
-                )
+            if is_spark_dataframe:
+                # For spark dataframe, only estimators ending with '_spark' are supported
+                estimator_list = [
+                    est for est in estimator_list if est.endswith("_spark")
+                ]
+                if len(estimator_list) == 0:
+                    raise ValueError(
+                        "Spark dataframes only support estimator names ending with `_spark`. Non-supported "
+                        "estimators are removed. No estimator is left."
+                    )
+                elif n_estimators != len(estimator_list):
+                    logger.warning(
+                        "Spark dataframes only support estimator names ending with `_spark`. Non-supported "
+                        "estimators are removed."
+                    )
+            else:
+                # For non-spark dataframe, only estimators not ending with '_spark' are supported
+                estimator_list = [
+                    est for est in estimator_list if not est.endswith("_spark")
+                ]
+                if len(estimator_list) == 0:
+                    raise ValueError(
+                        "Non-spark dataframes only support estimator names not ending with `_spark`. Non-supported "
+                        "estimators are removed. No estimator is left."
+                    )
+                elif n_estimators != len(estimator_list):
+                    logger.warning(
+                        "Non-spark dataframes only support estimator names not ending with `_spark`. Non-supported "
+                        "estimators are removed."
+                    )
+
+        if is_spark_dataframe and self._use_spark:
+            # For spark dataframe, use_spark must be False because spark models are trained in parallel themselves
+            self._use_spark = False
+            logger.warning(
+                "Spark dataframes support only spark.ml type models, which will be trained "
+                "with spark themselves, no need to start spark trials in flaml. "
+                "`use_spark` is set to False."
+            )
+
         # When no search budget is specified
         if no_budget:
             max_iter = len(estimator_list)
@@ -3866,7 +2713,7 @@ class AutoML(BaseEstimator):
                     [(estimator[0], estimator[1].params) for estimator in estimators]
                 )
             if len(estimators) > 1:
-                if self._state.task in CLASSIFICATION:
+                if self._state.task.is_classification():
                     from sklearn.ensemble import StackingClassifier as Stacker
                 else:
                     from sklearn.ensemble import StackingRegressor as Stacker
