+++ flaml/automl/logger.py
@@ -1,0 +1,7 @@
+import logging
+
+logger = logging.getLogger(__name__)
+logger_formatter = logging.Formatter(
+    "[%(name)s: %(asctime)s] {%(lineno)d} %(levelname)s - %(message)s", "%m-%d %H:%M:%S"
+)
+logger.propagate = False
