+++ flaml/default/suggest.py
@@ -1,11 +1,12 @@
+import os
 import numpy as np
 from sklearn.neighbors import NearestNeighbors
 import logging
 import pathlib
 import json
-import os
-from flaml.automl.data import CLASSIFICATION, DataTransformer
-from flaml.automl.ml import get_estimator_class, get_classification_objective
+from flaml.automl.data import DataTransformer
+from flaml.automl.task.task import CLASSIFICATION, get_classification_objective
+from flaml.automl.ml import get_estimator_class
 from flaml.version import __version__
 
 try:
@@ -67,6 +68,7 @@ def meta_feature(task, X_train, y_train, meta_feature_names):
 
 
 def load_config_predictor(estimator_name, task, location=None):
+    task = str(task)
     key = f"{location}/{estimator_name}/{task}"
     predictor = CONFIG_PREDICTORS.get(key)
     if predictor:
