+++ flaml/version.py
@@ -1,1 +1,1 @@
-__version__ = "1.1.3"
+__version__ = "1.2.0"
