+++ test/openai/test_notebook.py
@@ -38,5 +38,13 @@ def test_integrate_openai(save=False):
     run_notebook("integrate_openai.ipynb", save=save)
 
 
+@pytest.mark.skipif(
+    skip,
+    reason="do not run openai test if openai is not installed",
+)
+def test_integrate_chatgpt(save=False):
+    run_notebook("integrate_chatgpt_math.ipynb", save=save)
+
+
 if __name__ == "__main__":
-    test_integrate_openai(save=True)
+    test_integrate_chatgpt(save=True)
